refactor: remove unnecessary uint() casts in Day 17 tests

+++ day17_test.go
@@ -3,19 +3,19 @@ package adventofcode2018
 import "testing"
 
 func TestDay17Part1Example(t *testing.T) {
-	testWithParserBytes(t, 17, exampleFile, true, NewDay17, Day17, uint(57))
+	testWithParserBytes(t, 17, exampleFile, true, NewDay17, Day17, 57)
 }
 
 func TestDay17Part1(t *testing.T) {
-	testWithParserBytes(t, 17, file, true, NewDay17, Day17, uint(37858))
+	testWithParserBytes(t, 17, file, true, NewDay17, Day17, 37858)
 }
 
 func TestDay17Part2Example(t *testing.T) {
-	testWithParserBytes(t, 17, exampleFile, false, NewDay17, Day17, uint(29))
+	testWithParserBytes(t, 17, exampleFile, false, NewDay17, Day17, 29)
 }
 
 func TestDay17Part2(t *testing.T) {
-	testWithParserBytes(t, 17, file, false, NewDay17, Day17, uint(30410))
+	testWithParserBytes(t, 17, file, false, NewDay17, Day17, 30410)
 }
 
 func BenchmarkDay17Part1(b *testing.B) {
